// templates/xero/types.ts
export interface Address {
    AddressType: 'STREET' | 'POBOX';
    AddressLine1?: string;
    AddressLine2?: string;
    City: string;
    Region: string;
    PostalCode: string;
    Country: string;
}

export interface Phone {
    PhoneType: 'DEFAULT' | 'DDI' | 'FAX' | 'MOBILE';
    PhoneNumber: string;
    PhoneAreaCode: string;
    PhoneCountryCode: string;
}

export interface Contact {
    ContactID: string;
    ContactStatus: 'ACTIVE' | 'INACTIVE';
    ContactNumber?: string;
    Name: string;
    EmailAddress: string;
    BankAccountDetails: string;
    TaxNumber?: string;
    Addresses: Address[];
    Phones: Phone[];
    UpdatedDateUTC: string;
    ContactGroups: unknown[];
    IsSupplier: boolean;
    IsCustomer: boolean;
    SalesTrackingCategories: string[];
    PurchasesTrackingCategories: string[];
    ContactPersons: unknown[];
    HasValidationErrors: boolean;
    StatusAttributeString: string;
}

interface SlimItem {
    ItemID: string;
    Name: string;
    Code: string;
}

interface Tracking {
    TrackingCategoryID: string;
    Name: string;
    Option: string;
}
// ...
export interface XeroTrackingCategory {
    Name: string;
    Option: string;
    TrackingCategoryID: string;
    TrackingOptionID: string;
    Options: string[];
}

export interface XeroJournalLine {
    JournalLineID: string;
    AccountID: string;
    AccountCode: string;
    AccountType: string;
    AccountName: string;
    Description?: string;
    NetAmount: number;
    GrossAmount: number;
    TaxAmount: number;
    TaxType?: string;
    TaxName?: string;
    TrackingCategories: XeroTrackingCategory[];
}

export interface XeroBankTransaction {
    BankTransactionID: string;
    Type: 'SPEND' | 'RECEIVE' | 'RECEIVE-OVERPAYMENT' | 'RECEIVE-PREPAYMENT' | 'SPEND-OVERPAYMENT' | 'SPEND-PREPAYMENT' | 'RECEIVE-TRANSFER' | 'SPEND-TRANSFER';
    BankAccount: {
        AccountID: string;
        Code: string;
        Name: string;
    };
    Contact?: {
        ContactID: string;
        Name: string;
    };
    Date: string;
    DateString: string;
    Status: 'AUTHORISED' | 'DELETED';
    Reference?: string;
    IsReconciled: boolean;
    CurrencyCode: string;
    CurrencyRate?: number;
    Total: number;
    SubTotal: number;
    TotalTax: number;
    LineAmountTypes: 'Inclusive' | 'Exclusive' | 'NoTax';
    LineItems: XeroBankTransactionLineItem[];
    UpdatedDateUTC: string;
    Url?: string;
    HasAttachments: boolean;
}

export interface XeroBankTransactionLineItem {
    Description: string;
    Quantity: number;
    UnitAmount: number;
    AccountCode: string;
    ItemCode?: string;
    LineItemID: string;
    TaxType?: string;
    TaxAmount: number;
    LineAmount: number;
    Tracking?: XeroTrackingCategory[];
    AccountId: string;
}